pip-requirements.txt
python-gitlab==2.0.1